[ALPS05758892] imgsys: 64 bit iova
64 bit iova

MTK-Commit-Id: a02883e846d0f8b8cc62ff8bb3aa7fc955bd38d2

CR-Id: ALPS05758892
Feature: [Android Default] Camera Application Basic Functions

+++ drivers/media/platform/mtk-isp/imgsys/mtk_imgsys-dev.c
@@ -540,7 +540,7 @@ static u64 mtk_imgsys_get_iova(struct dma_buf *dma_buf, s32 ionFd,
 
 	if (cache) {
 		pr_debug("%s cache hit\n", __func__);
-		return (u32)iova_info->dma_addr;
+		return iova_info->dma_addr;
 	}
 
 	dev = imgsys_dev->dev;
